long instead int added extra lines deleted

+++ src/main/java/Main.java
@@ -3,7 +3,7 @@ import ru.netology.stats.StatsService;
 public class Main {
     public static void main(String[] args) {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         System.out.println(service.getAllSales(sales));
         System.out.println(service.getAverageSalesAmount(sales));
         System.out.println(service.getMaxSales(sales));
